Calculate Grantham distance between HLA alleles

hlaCallsGranthamDistance calculate Grantham distance between two HLA alleles of a given, using original formula by Grantham R. 1974..

hlaCallsGranthamDistance(
  hla_calls,
  genes = c("A", "B", "C"),
  aa_selection = "binding_groove"
)

Arguments

hla_calls
HLA calls data frame, as returned by readHlaCalls function.

genes
Character vector specifying genes for which allelic distance should be calculated.

aa_selection
String specifying variable region in peptide binding groove which should be considered for Grantham distance calculation. Valid choices includes: "binding_groove", "B_pocket", "F_pocket". See details for more information.

Value

Data frame of normalized Grantham distances between pairs of alleles for each specified HLA gene. First column (ID) is the same as in hla_calls, further columns are named as given by genes.

Details

Grantham distance is calculated only for class I HLA alleles. First exons forming the variable region in the peptide binding groove are selected. Here we provide option to choose either "binding_groove" - exon 2 and 3 (positions 1-182 in IMGT/HLA alignments, however here we take 2-182 as many 1st positions are missing), "B_pocket" - residues 7, 9, 24, 25, 34, 45, 63, 66, 67, 70, 99 and "F_pocket" - residues 77, 80, 81, 84, 95, 116, 123, 143, 146, 147. Then all the alleles containing gaps, stop codons or indels are discarded. Finally distance is calculated for each pair.

See Robinson J. 2017. for more details on the choice of exons 2 and 3.

Examples

hlaCallsGranthamDistance(MiDAS_tut_HLA, genes = "A")

#>        ID          A
#> 1    C001  3.8121547
#> 2    C002  0.8729282
#> 3    C003  3.5580110
#> 4    C004 10.6077348
#> 5    C005  9.1712707
#> 6    C006  5.7458564
#> 7    C007 10.5027624
#> 8    C008  8.1049724
#> 9    C009 11.3867403
#> 10   C010  7.4585635
#> 11   C011 10.5027624
#> 12   C012  0.0000000
#> 13   C013 10.7237569
#> 14   C014  8.8563536
#> 15   C015  6.4254144
#> 16   C016  7.0718232
#> 17   C017  3.0828729
#> 18   C018  8.1049724
#> 19   C019  6.5635359
#> 20   C020  6.6408840
#> 21   C021  8.1049724
#> 22   C022  9.1712707
#> 23   C023  9.9116022
#> 24   C024  9.9226519
#> 25   C025  4.9668508
#> 26   C026 10.7900552
#> 27   C027  0.0000000
#> 28   C028 10.5027624
#> 29   C029  0.0000000
#> 30   C030  4.4585635
#> 31   C031  7.0718232
#> 32   C032 10.6574586
#> 33   C033 10.5027624
#> 34   C034  5.2044199
#> 35   C035 10.5027624
#> 36   C036 10.5966851
#> 37   C037  0.0000000
#> 38   C038 10.9447514
#> 39   C039  8.9447514
#> 40   C040  6.1270718
#> 41   C041  6.3204420
#> 42   C042  6.4254144
#> 43   C043 10.5027624
#> 44   C044 13.0441989
#> 45   C045  7.6022099
#> 46   C046  9.1712707
#> 47   C047  3.0828729
#> 48   C048  7.0386740
#> 49   C049 10.3425414
#> 50   C050  3.0828729
#> 51   C051  9.1215470
#> 52   C052 12.4861878
#> 53   C053  0.0000000
#> 54   C054  5.7458564
#> 55   C055 10.6077348
#> 56   C056  0.0000000
#> 57   C057  2.2209945
#> 58   C058 10.6629834
#> 59   C059 10.9447514
#> 60   C060 10.5027624
#> 61   C061  6.4254144
#> 62   C062  9.9116022
#> 63   C063  8.9447514
#> 64   C064  0.0000000
#> 65   C065  0.0000000
#> 66   C066 10.2928177
#> 67   C067  8.9723757
#> 68   C068  8.1878453
#> 69   C069  9.1712707
#> 70   C070 10.8232044
#> 71   C071  0.0000000
#> 72   C072 10.6077348
#> 73   C073  4.3204420
#> 74   C074 10.2928177
#> 75   C075  5.1215470
#> 76   C076  7.4198895
#> 77   C077  8.9171271
#> 78   C078  5.3591160
#> 79   C079  0.0000000
#> 80   C080  8.9447514
#> 81   C081  6.4254144
#> 82   C082  2.8176796
#> 83   C083  0.0000000
#> 84   C084  8.5911602
#> 85   C085  8.5801105
#> 86   C086  8.9723757
#> 87   C087  6.4254144
#> 88   C088 10.6574586
#> 89   C089  4.1823204
#> 90   C090  0.0000000
#> 91   C091 10.5027624
#> 92   C092 10.4861878
#> 93   C093 12.1325967
#> 94   C094 12.9502762
#> 95   C095  1.4861878
#> 96   C096  7.4198895
#> 97   C097  5.2044199
#> 98   C098 10.2928177
#> 99   C099 13.6961326
#> 100  C100  3.0718232
#> 101  C101 10.4861878
#> 102  C102 10.5027624
#> 103  C103  0.0000000
#> 104  C104  8.9723757
#> 105  C105  6.4254144
#> 106  C106  7.0883978
#> 107  C107  4.8232044
#> 108  C108  7.4198895
#> 109  C109 12.8674033
#> 110  C110  5.3038674
#> 111  C111 10.6574586
#> 112  C112 10.5027624
#> 113  C113  7.4585635
#> 114  C114  9.1712707
#> 115  C115  7.4475138
#> 116  C116  0.0000000
#> 117  C117  0.0000000
#> 118  C118  0.0000000
#> 119  C119  0.1215470
#> 120  C120  4.1823204
#> 121  C121 10.5966851
#> 122  C122  3.0828729
#> 123  C123  0.8729282
#> 124  C124 13.0441989
#> 125  C125  4.7569061
#> 126  C126  3.0828729
#> 127  C127 10.9447514
#> 128  C128  6.2983425
#> 129  C129  5.0441989
#> 130  C130  3.2044199
#> 131  C131  6.3204420
#> 132  C132  5.7458564
#> 133  C133  6.4254144
#> 134  C134 10.5027624
#> 135  C135 10.6961326
#> 136  C136 10.2486188
#> 137  C137  3.7790055
#> 138  C138  8.9723757
#> 139  C139  2.3259669
#> 140  C140  7.2928177
#> 141  C141 10.5027624
#> 142  C142 10.6629834
#> 143  C143  6.4254144
#> 144  C144  3.0718232
#> 145  C145 11.6685083
#> 146  C146  7.0386740
#> 147  C147 10.5027624
#> 148  C148 10.5027624
#> 149  C149  9.4033149
#> 150  C150  8.3038674
#> 151  C151  0.0000000
#> 152  C152  3.2044199
#> 153  C153  4.2651934
#> 154  C154 10.2928177
#> 155  C155  0.8729282
#> 156  C156 10.6077348
#> 157  C157 10.8011050
#> 158  C158 10.2596685
#> 159  C159  8.1049724
#> 160  C160  7.0386740
#> 161  C161  7.8011050
#> 162  C162  5.2044199
#> 163  C163 10.2928177
#> 164  C164  6.4254144
#> 165  C165  0.0000000
#> 166  C166  6.4254144
#> 167  C167  0.0000000
#> 168  C168 10.6961326
#> 169  C169  0.0000000
#> 170  C170  8.2375691
#> 171  C171  9.5911602
#> 172  C172 10.2928177
#> 173  C173  0.0000000
#> 174  C174  7.5193370
#> 175  C175  7.6022099
#> 176  C176 10.1988950
#> 177  C177  0.0000000
#> 178  C178  5.7458564
#> 179  C179  0.0000000
#> 180  C180  0.0000000
#> 181  C181 10.2928177
#> 182  C182 10.9447514
#> 183  C183  9.7182320
#> 184  C184  9.0773481
#> 185  C185  6.4254144
#> 186  C186  8.9447514
#> 187  C187 13.0441989
#> 188  C188  0.0000000
#> 189  C189  2.7016575
#> 190  C190 12.9502762
#> 191  C191  3.5469613
#> 192  C192  0.0000000
#> 193  C193 10.4861878
#> 194  C194 10.4861878
#> 195  C195  5.4364641
#> 196  C196  6.6408840
#> 197  C197 10.5027624
#> 198  C198 13.0441989
#> 199  C199  5.7458564
#> 200  C200 11.3867403
#> 201  C201  6.4254144
#> 202  C202 10.4861878
#> 203  C203  7.6243094
#> 204  C204  8.8563536
#> 205  C205 10.5027624
#> 206  C206  5.9613260
#> 207  C207 10.5027624
#> 208  C208  8.1878453
#> 209  C209  3.2596685
#> 210  C210  5.0883978
#> 211  C211  7.4198895
#> 212  C212 10.7955801
#> 213  C213  0.0000000
#> 214  C214  6.0055249
#> 215  C215  9.1767956
#> 216  C216  6.6408840
#> 217  C217  6.4254144
#> 218  C218 10.6574586
#> 219  C219 10.9447514
#> 220  C220  8.1049724
#> 221  C221  3.2044199
#> 222  C222  5.7458564
#> 223  C223  8.1049724
#> 224  C224 12.4861878
#> 225  C225  0.0000000
#> 226  C226  5.7458564
#> 227  C227  6.2320442
#> 228  C228  3.5469613
#> 229  C229  4.6740331
#> 230  C230  0.0000000
#> 231  C231  9.7403315
#> 232  C232 10.6574586
#> 233  C233 10.5966851
#> 234  C234  0.0000000
#> 235  C235  9.0773481
#> 236  C236  7.4033149
#> 237  C237  0.0000000
#> 238  C238  9.9116022
#> 239  C239 10.7237569
#> 240  C240 13.0441989
#> 241  C241  6.4254144
#> 242  C242  6.2320442
#> 243  C243 10.3314917
#> 244  C244 12.4861878
#> 245  C245  5.1215470
#> 246  C246  0.0000000
#> 247  C247  8.1049724
#> 248  C248  7.0718232
#> 249  C249  5.4696133
#> 250  C250  0.0000000
#> 251  C251  3.7900552
#> 252  C252  0.0000000
#> 253  C253 10.6464088
#> 254  C254  2.3977901
#> 255  C255  8.2375691
#> 256  C256 11.3867403
#> 257  C257 10.2596685
#> 258  C258  6.2983425
#> 259  C259 10.6961326
#> 260  C260  6.9613260
#> 261  C261 10.6574586
#> 262  C262  0.0000000
#> 263  C263  7.0718232
#> 264  C264  6.5138122
#> 265  C265 11.0386740
#> 266  C266  8.9447514
#> 267  C267 10.2928177
#> 268  C268  7.4198895
#> 269  C269  4.2651934
#> 270  C270  5.3591160
#> 271  C271  3.0718232
#> 272  C272 10.9447514
#> 273  C273 10.9447514
#> 274  C274  4.8232044
#> 275  C275  6.3812155
#> 276  C276  6.3204420
#> 277  C277 10.2596685
#> 278  C278  3.2596685
#> 279  C279  6.3204420
#> 280  C280  8.1878453
#> 281  C281  8.1878453
#> 282  C282 10.2928177
#> 283  C283 10.5027624
#> 284  C284  9.7237569
#> 285  C285  9.1712707
#> 286  C286 10.5027624
#> 287  C287 10.5027624
#> 288  C288  6.4254144
#> 289  C289  8.9447514
#> 290  C290 10.5027624
#> 291  C291  0.0000000
#> 292  C292  1.2928177
#> 293  C293  0.0000000
#> 294  C294  8.1878453
#> 295  C295  6.4254144
#> 296  C296  8.9447514
#> 297  C297  5.7458564
#> 298  C298  8.9944751
#> 299  C299 10.7513812
#> 300  C300 13.0441989
#> 301  C301  5.1215470
#> 302  C302  6.4254144
#> 303  C303 10.6077348
#> 304  C304  5.2044199
#> 305  C305  6.4254144
#> 306  C306 10.9447514
#> 307  C307  8.1049724
#> 308  C308  0.0000000
#> 309  C309  8.2375691
#> 310  C310  3.0718232
#> 311  C311  8.9723757
#> 312  C312 12.1325967
#> 313  C313 10.6077348
#> 314  C314  1.0773481
#> 315  C315  3.5469613
#> 316  C316 10.6574586
#> 317  C317 10.9447514
#> 318  C318 10.5027624
#> 319  C319 13.0441989
#> 320  C320  7.9613260
#> 321  C321 10.6574586
#> 322  C322  5.7458564
#> 323  C323  8.2375691
#> 324  C324  7.6022099
#> 325  C325  3.0828729
#> 326  C326  0.3093923
#> 327  C327 10.5027624
#> 328  C328  8.3038674
#> 329  C329  6.6408840
#> 330  C330  5.3922652
#> 331  C331 10.6077348
#> 332  C332  3.7071823
#> 333  C333 10.5027624
#> 334  C334  6.1270718
#> 335  C335  3.7845304
#> 336  C336  8.9171271
#> 337  C337  6.4254144
#> 338  C338 10.5248619
#> 339  C339  5.5303867
#> 340  C340  5.3591160
#> 341  C341  4.1325967
#> 342  C342 10.9889503
#> 343  C343  9.2651934
#> 344  C344  8.9723757
#> 345  C345 10.7955801
#> 346  C346  5.7458564
#> 347  C347  7.9171271
#> 348  C348 10.5027624
#> 349  C349  5.4254144
#> 350  C350 10.4861878
#> 351  C351  5.4364641
#> 352  C352  7.4475138
#> 353  C353  6.6408840
#> 354  C354  6.6408840
#> 355  C355  9.9116022
#> 356  C356 10.5027624
#> 357  C357  3.2044199
#> 358  C358  8.5966851
#> 359  C359  6.4254144
#> 360  C360 10.5027624
#> 361  C361  3.2044199
#> 362  C362  9.3922652
#> 363  C363 10.4861878
#> 364  C364 10.4861878
#> 365  C365  4.9060773
#> 366  C366 10.6077348
#> 367  C367  8.9447514
#> 368  C368  3.1215470
#> 369  C369  5.2983425
#> 370  C370  7.2707182
#> 371  C371 12.9502762
#> 372  C372  5.2044199
#> 373  C373  7.4033149
#> 374  C374  3.7790055
#> 375  C375  8.9447514
#> 376  C376  6.2541436
#> 377  C377  0.0000000
#> 378  C378  8.9723757
#> 379  C379  8.9723757
#> 380  C380  9.1712707
#> 381  C381 10.6574586
#> 382  C382  5.3591160
#> 383  C383 10.6077348
#> 384  C384 10.6574586
#> 385  C385 10.5027624
#> 386  C386  6.4254144
#> 387  C387 10.6574586
#> 388  C388  6.4254144
#> 389  C389  8.1823204
#> 390  C390 10.2596685
#> 391  C391  6.4254144
#> 392  C392  3.2044199
#> 393  C393 11.3425414
#> 394  C394  8.9171271
#> 395  C395 12.4861878
#> 396  C396  6.2320442
#> 397  C397 10.6574586
#> 398  C398  7.5193370
#> 399  C399 10.6574586
#> 400  C400 10.6574586
#> 401  C401 10.6574586
#> 402  C402  9.4033149
#> 403  C403  0.0000000
#> 404  C404 10.9447514
#> 405  C405 10.6574586
#> 406  C406  0.0000000
#> 407  C407  0.0000000
#> 408  C408  7.0994475
#> 409  C409  4.8287293
#> 410  C410  8.1049724
#> 411  C411  7.5193370
#> 412  C412 10.1988950
#> 413  C413  4.8508287
#> 414  C414 13.0441989
#> 415  C415 10.6574586
#> 416  C416 10.6077348
#> 417  C417  5.3591160
#> 418  C418  3.7790055
#> 419  C419  9.0773481
#> 420  C420  3.7790055
#> 421  C421  6.9889503
#> 422  C422  3.2044199
#> 423  C423  4.0607735
#> 424  C424 10.2928177
#> 425  C425  6.2983425
#> 426  C426  1.9116022
#> 427  C427 10.4861878
#> 428  C428  0.0000000
#> 429  C429  2.7845304
#> 430  C430 10.9447514
#> 431  C431  6.4143646
#> 432  C432 11.6685083
#> 433  C433  6.4254144
#> 434  C434  9.7403315
#> 435  C435  3.0828729
#> 436  C436 11.3867403
#> 437  C437  0.0000000
#> 438  C438 13.0441989
#> 439  C439 10.6961326
#> 440  C440  4.6740331
#> 441  C441  5.2044199
#> 442  C442  0.0000000
#> 443  C443  6.4254144
#> 444  C444  6.4254144
#> 445  C445  8.1049724
#> 446  C446  8.9447514
#> 447  C447  8.9723757
#> 448  C448  5.9723757
#> 449  C449 10.3425414
#> 450  C450  1.4475138
#> 451  C451  6.4254144
#> 452  C452  8.1049724
#> 453  C453  0.0000000
#> 454  C454  6.4254144
#> 455  C455  4.1325967
#> 456  C456 10.6574586
#> 457  C457 10.5027624
#> 458  C458 10.5027624
#> 459  C459  5.2044199
#> 460  C460  8.1049724
#> 461  C461  8.5966851
#> 462  C462 10.5027624
#> 463  C463  0.7513812
#> 464  C464 10.6574586
#> 465  C465  8.9723757
#> 466  C466  0.0000000
#> 467  C467 10.6961326
#> 468  C468 10.5248619
#> 469  C469 10.5027624
#> 470  C470  7.5193370
#> 471  C471 12.4861878
#> 472  C472 10.6574586
#> 473  C473 10.5027624
#> 474  C474  8.2375691
#> 475  C475 10.6574586
#> 476  C476  3.0828729
#> 477  C477 10.5027624
#> 478  C478  0.0000000
#> 479  C479  8.9723757
#> 480  C480 10.6574586
#> 481  C481  0.8729282
#> 482  C482  0.0000000
#> 483  C483  9.7403315
#> 484  C484  6.4254144
#> 485  C485 10.2928177
#> 486  C486  6.4254144
#> 487  C487 10.5027624
#> 488  C488 10.6464088
#> 489  C489  2.8176796
#> 490  C490  3.0718232
#> 491  C491 10.5027624
#> 492  C492  3.2044199
#> 493  C493  0.0000000
#> 494  C494  8.9944751
#> 495  C495 10.6077348
#> 496  C496  0.0000000
#> 497  C497 10.2928177
#> 498  C498 11.2707182
#> 499  C499  9.2265193
#> 500  C500  6.3204420
#> 501  P001 10.9447514
#> 502  P002  0.0000000
#> 503  P003  8.1049724
#> 504  P004 10.6629834
#> 505  P005 10.6574586
#> 506  P006  5.7458564
#> 507  P007  6.4254144
#> 508  P008  0.0000000
#> 509  P009  7.0386740
#> 510  P010  3.0718232
#> 511  P011  8.0441989
#> 512  P012  0.0000000
#> 513  P013  0.0000000
#> 514  P014  7.5193370
#> 515  P015 10.6574586
#> 516  P016  8.9447514
#> 517  P017  4.4309392
#> 518  P018 10.6574586
#> 519  P019 10.6574586
#> 520  P020 10.6077348
#> 521  P021  8.1049724
#> 522  P022  8.9723757
#> 523  P023  1.2928177
#> 524  P024  3.0718232
#> 525  P025 10.3314917
#> 526  P026  8.1049724
#> 527  P027  5.7458564
#> 528  P028  2.1436464
#> 529  P029  0.0000000
#> 530  P030  7.6243094
#> 531  P031 10.6961326
#> 532  P032  7.0386740
#> 533  P033  8.5911602
#> 534  P034 10.9447514
#> 535  P035  6.4254144
#> 536  P036  6.4254144
#> 537  P037 10.6574586
#> 538  P038  0.0000000
#> 539  P039  6.4254144
#> 540  P040 10.1988950
#> 541  P041 10.6574586
#> 542  P042 10.6574586
#> 543  P043  0.0000000
#> 544  P044 10.6574586
#> 545  P045  6.5138122
#> 546  P046  7.6243094
#> 547  P047  4.6740331
#> 548  P048 10.5027624
#> 549  P049  7.8618785
#> 550  P050  4.6740331
#> 551  P051  6.4254144
#> 552  P052  0.0000000
#> 553  P053  6.1270718
#> 554  P054  8.9723757
#> 555  P055  3.2596685
#> 556  P056  5.0441989
#> 557  P057 10.6574586
#> 558  P058 12.4861878
#> 559  P059  8.9723757
#> 560  P060 10.5027624
#> 561  P061 10.6077348
#> 562  P062  8.1049724
#> 563  P063  3.7900552
#> 564  P064  0.0000000
#> 565  P065  0.0000000
#> 566  P066  5.8397790
#> 567  P067 10.5027624
#> 568  P068  8.9171271
#> 569  P069  6.3204420
#> 570  P070  0.0000000
#> 571  P071 10.6574586
#> 572  P072  0.0000000
#> 573  P073 10.9447514
#> 574  P074  6.5138122
#> 575  P075  3.0828729
#> 576  P076  0.0000000
#> 577  P077  8.9723757
#> 578  P078  6.1270718
#> 579  P079 13.0441989
#> 580  P080 10.5027624
#> 581  P081 10.2430939
#> 582  P082 10.4861878
#> 583  P083  6.6408840
#> 584  P084 10.6574586
#> 585  P085 10.5027624
#> 586  P086  5.0441989
#> 587  P087  8.1049724
#> 588  P088  5.3038674
#> 589  P089  8.9723757
#> 590  P090  8.6795580
#> 591  P091  0.0000000
#> 592  P092 10.6629834
#> 593  P093 10.5027624
#> 594  P094 10.9447514
#> 595  P095 10.6574586
#> 596  P096 10.6574586
#> 597  P097  3.0828729
#> 598  P098  3.7845304
#> 599  P099  3.7790055
#> 600  P100  8.9447514
#> 601  P101 10.6077348
#> 602  P102  8.8563536
#> 603  P103  7.6243094
#> 604  P104 12.9502762
#> 605  P105  0.0000000
#> 606  P106 10.5027624
#> 607  P107  5.7458564
#> 608  P108  3.7790055
#> 609  P109  8.1049724
#> 610  P110  0.0000000
#> 611  P111 10.6574586
#> 612  P112 10.5027624
#> 613  P113  1.2928177
#> 614  P114 10.7955801
#> 615  P115 10.6795580
#> 616  P116  4.8011050
#> 617  P117  0.0000000
#> 618  P118 10.5027624
#> 619  P119  3.7900552
#> 620  P120  7.9834254
#> 621  P121  3.0718232
#> 622  P122 13.0441989
#> 623  P123  5.3591160
#> 624  P124  5.4972376
#> 625  P125 10.6574586
#> 626  P126  0.0000000
#> 627  P127 11.6685083
#> 628  P128  6.4254144
#> 629  P129 10.6574586
#> 630  P130  8.9723757
#> 631  P131  0.0000000
#> 632  P132 10.5027624
#> 633  P133 10.3425414
#> 634  P134 10.6961326
#> 635  P135  0.0000000
#> 636  P136 13.0441989
#> 637  P137 10.2596685
#> 638  P138 10.6629834
#> 639  P139  1.3259669
#> 640  P140  5.7458564
#> 641  P141 10.6077348
#> 642  P142 10.4861878
#> 643  P143  0.0000000
#> 644  P144 10.5966851
#> 645  P145 12.5303867
#> 646  P146  7.5193370
#> 647  P147 10.5027624
#> 648  P148  8.1049724
#> 649  P149  8.2375691
#> 650  P150  0.0000000
#> 651  P151  6.4254144
#> 652  P152 10.9447514
#> 653  P153  0.0000000
#> 654  P154  6.4254144
#> 655  P155 10.5027624
#> 656  P156  0.0000000
#> 657  P157  8.9447514
#> 658  P158 10.6077348
#> 659  P159 10.2596685
#> 660  P160 10.6961326
#> 661  P161 10.6574586
#> 662  P162  7.4198895
#> 663  P163  3.2044199
#> 664  P164  3.8121547
#> 665  P165  4.1823204
#> 666  P166  5.2044199
#> 667  P167  7.8618785
#> 668  P168  0.0000000
#> 669  P169  9.1712707
#> 670  P170  1.1602210
#> 671  P171  0.1325967
#> 672  P172  4.1712707
#> 673  P173  9.0773481
#> 674  P174  9.7403315
#> 675  P175 10.6077348
#> 676  P176 10.6574586
#> 677  P177  4.6740331
#> 678  P178 10.5027624
#> 679  P179  3.5469613
#> 680  P180  9.4585635
#> 681  P181  5.7458564
#> 682  P182  6.4254144
#> 683  P183 10.5027624
#> 684  P184 11.7403315
#> 685  P185 10.6077348
#> 686  P186 10.6961326
#> 687  P187 10.5027624
#> 688  P188 10.6574586
#> 689  P189  0.0000000
#> 690  P190  9.2265193
#> 691  P191 10.6574586
#> 692  P192  7.9834254
#> 693  P193  0.0000000
#> 694  P194 10.6077348
#> 695  P195 10.7955801
#> 696  P196  0.0000000
#> 697  P197 10.5027624
#> 698  P198 10.6574586
#> 699  P199  6.3204420
#> 700  P200 13.0441989
#> 701  P201  6.5138122
#> 702  P202 10.6961326
#> 703  P203  8.0497238
#> 704  P204 10.3314917
#> 705  P205  4.2707182
#> 706  P206  3.0828729
#> 707  P207  0.0000000
#> 708  P208  5.7458564
#> 709  P209 10.4861878
#> 710  P210  0.0000000
#> 711  P211  3.7790055
#> 712  P212  0.0000000
#> 713  P213 10.5027624
#> 714  P214  6.4254144
#> 715  P215  0.0000000
#> 716  P216  3.8287293
#> 717  P217 12.0220994
#> 718  P218  8.1049724
#> 719  P219  4.6740331
#> 720  P220  9.7182320
#> 721  P221  6.4254144
#> 722  P222 10.5027624
#> 723  P223 10.8895028
#> 724  P224  8.9447514
#> 725  P225  0.0000000
#> 726  P226  0.0000000
#> 727  P227 10.8011050
#> 728  P228  4.6740331
#> 729  P229 10.5027624
#> 730  P230  0.0000000
#> 731  P231  0.0000000
#> 732  P232  5.9723757
#> 733  P233 10.5248619
#> 734  P234  7.5193370
#> 735  P235  9.6408840
#> 736  P236 10.2596685
#> 737  P237  0.0000000
#> 738  P238 10.6574586
#> 739  P239 10.7292818
#> 740  P240 10.5027624
#> 741  P241 10.9889503
#> 742  P242 12.4861878
#> 743  P243  3.5469613
#> 744  P244  3.1657459
#> 745  P245 10.6574586
#> 746  P246  6.3812155
#> 747  P247  4.1823204
#> 748  P248  6.4254144
#> 749  P249  6.4254144
#> 750  P250  8.9723757
#> 751  P251  7.5193370
#> 752  P252  7.7292818
#> 753  P253  0.0000000
#> 754  P254  8.2375691
#> 755  P255  0.0000000
#> 756  P256  6.4254144
#> 757  P257  0.0000000
#> 758  P258  0.0000000
#> 759  P259 10.5027624
#> 760  P260  8.8342541
#> 761  P261  8.9447514
#> 762  P262 10.5027624
#> 763  P263 10.5027624
#> 764  P264  8.9723757
#> 765  P265  6.3812155
#> 766  P266 10.6629834
#> 767  P267  8.9723757
#> 768  P268  7.9613260
#> 769  P269  5.7458564
#> 770  P270  8.9171271
#> 771  P271 10.6077348
#> 772  P272 10.5027624
#> 773  P273 10.5027624
#> 774  P274 10.5027624
#> 775  P275  6.4254144
#> 776  P276 10.2928177
#> 777  P277  6.4254144
#> 778  P278  0.0000000
#> 779  P279  3.0828729
#> 780  P280  8.9723757
#> 781  P281  0.0000000
#> 782  P282 10.5027624
#> 783  P283 10.6077348
#> 784  P284 10.6961326
#> 785  P285 10.1988950
#> 786  P286 10.6961326
#> 787  P287 10.4861878
#> 788  P288 10.9447514
#> 789  P289  8.1049724
#> 790  P290  7.4198895
#> 791  P291  7.0386740
#> 792  P292  4.6740331
#> 793  P293  6.0055249
#> 794  P294 10.1988950
#> 795  P295  7.0386740
#> 796  P296  3.5469613
#> 797  P297 12.4861878
#> 798  P298  7.4475138
#> 799  P299  3.0718232
#> 800  P300 10.9447514
#> 801  P301  4.6077348
#> 802  P302 12.4861878
#> 803  P303  5.2044199
#> 804  P304  3.2044199
#> 805  P305 10.2430939
#> 806  P306  5.3591160
#> 807  P307  8.9723757
#> 808  P308 10.2928177
#> 809  P309  0.0000000
#> 810  P310  8.8563536
#> 811  P311  0.0000000
#> 812  P312 10.5027624
#> 813  P313 10.3149171
#> 814  P314  4.9668508
#> 815  P315  8.9447514
#> 816  P316  9.6574586
#> 817  P317  8.1049724
#> 818  P318  1.3259669
#> 819  P319  8.8342541
#> 820  P320  9.1712707
#> 821  P321 10.6961326
#> 822  P322  6.0220994
#> 823  P323  3.0828729
#> 824  P324 10.6961326
#> 825  P325  6.3204420
#> 826  P326  0.0000000
#> 827  P327  6.4254144
#> 828  P328  6.1270718
#> 829  P329  5.2044199
#> 830  P330 10.2928177
#> 831  P331 10.6574586
#> 832  P332  7.4198895
#> 833  P333 10.5027624
#> 834  P334  1.1933702
#> 835  P335  7.4198895
#> 836  P336  9.1215470
#> 837  P337  4.9668508
#> 838  P338  7.9834254
#> 839  P339  0.8729282
#> 840  P340  3.0718232
#> 841  P341 10.6574586
#> 842  P342 10.5966851
#> 843  P343 10.6077348
#> 844  P344  7.1823204
#> 845  P345  5.9723757
#> 846  P346 10.2928177
#> 847  P347 10.5027624
#> 848  P348  4.0607735
#> 849  P349  4.4419890
#> 850  P350 12.5303867
#> 851  P351  9.5856354
#> 852  P352  7.0718232
#> 853  P353  0.0000000
#> 854  P354 10.5027624
#> 855  P355  8.9723757
#> 856  P356 10.6961326
#> 857  P357  5.1215470
#> 858  P358  8.2375691
#> 859  P359 10.5248619
#> 860  P360  5.7458564
#> 861  P361  8.1049724
#> 862  P362  7.0718232
#> 863  P363  8.0497238
#> 864  P364 10.5027624
#> 865  P365  4.7569061
#> 866  P366  4.6740331
#> 867  P367  3.0828729
#> 868  P368  9.1712707
#> 869  P369  8.8176796
#> 870  P370 10.6574586
#> 871  P371  0.0000000
#> 872  P372  8.9447514
#> 873  P373  2.8176796
#> 874  P374 11.9668508
#> 875  P375 10.6574586
#> 876  P376 10.2928177
#> 877  P377  6.3812155
#> 878  P378  0.0000000
#> 879  P379  7.7292818
#> 880  P380  4.4419890
#> 881  P381  7.9834254
#> 882  P382  5.0441989
#> 883  P383  3.3259669
#> 884  P384 10.5027624
#> 885  P385  3.4088398
#> 886  P386  7.5193370
#> 887  P387  0.0000000
#> 888  P388  6.3204420
#> 889  P389 10.6574586
#> 890  P390  7.7292818
#> 891  P391  4.1823204
#> 892  P392  8.9447514
#> 893  P393  5.7458564
#> 894  P394 10.9447514
#> 895  P395  8.9171271
#> 896  P396  0.0000000
#> 897  P397  5.7458564
#> 898  P398  0.0000000
#> 899  P399  5.9613260
#> 900  P400 10.9447514
#> 901  P401  3.5469613
#> 902  P402 10.5027624
#> 903  P403  3.0718232
#> 904  P404  9.7403315
#> 905  P405  4.2651934
#> 906  P406  1.0773481
#> 907  P407  3.0718232
#> 908  P408  6.4254144
#> 909  P409  3.7900552
#> 910  P410 10.5027624
#> 911  P411  6.7845304
#> 912  P412  8.9447514
#> 913  P413  3.0828729
#> 914  P414  3.3259669
#> 915  P415  7.4585635
#> 916  P416 10.5027624
#> 917  P417 12.4861878
#> 918  P418 10.5027624
#> 919  P419  5.7458564
#> 920  P420  5.7458564
#> 921  P421  8.9171271
#> 922  P422 10.6574586
#> 923  P423 10.6961326
#> 924  P424  7.4198895
#> 925  P425  0.0000000
#> 926  P426  4.3204420
#> 927  P427  2.8176796
#> 928  P428 10.6574586
#> 929  P429  8.9723757
#> 930  P430  8.9723757
#> 931  P431 10.6077348
#> 932  P432  3.5469613
#> 933  P433 10.5248619
#> 934  P434  3.8618785
#> 935  P435 10.2928177
#> 936  P436  6.4254144
#> 937  P437  8.1878453
#> 938  P438  8.9447514
#> 939  P439  5.7458564
#> 940  P440  9.9116022
#> 941  P441  5.3591160
#> 942  P442 12.4861878
#> 943  P443  8.2375691
#> 944  P444  9.6408840
#> 945  P445 11.9668508
#> 946  P446  3.0828729
#> 947  P447  8.9944751
#> 948  P448  4.6740331
#> 949  P449  3.0718232
#> 950  P450 10.5966851
#> 951  P451 10.6574586
#> 952  P452 10.6243094
#> 953  P453 10.6961326
#> 954  P454 10.6961326
#> 955  P455  6.6408840
#> 956  P456  3.7845304
#> 957  P457 10.6961326
#> 958  P458  0.0000000
#> 959  P459  7.0386740
#> 960  P460  2.6187845
#> 961  P461 10.5027624
#> 962  P462 10.8011050
#> 963  P463  8.9447514
#> 964  P464  0.0000000
#> 965  P465  9.4309392
#> 966  P466  5.3259669
#> 967  P467  6.4254144
#> 968  P468  7.2486188
#> 969  P469 10.6077348
#> 970  P470  8.8563536
#> 971  P471  0.8729282
#> 972  P472  7.0718232
#> 973  P473  8.1049724
#> 974  P474  0.0000000
#> 975  P475  7.5193370
#> 976  P476  8.1049724
#> 977  P477  7.8618785
#> 978  P478 10.5027624
#> 979  P479 10.6077348
#> 980  P480  3.0828729
#> 981  P481  3.2044199
#> 982  P482  0.0000000
#> 983  P483  6.9116022
#> 984  P484 10.5027624
#> 985  P485  5.0441989
#> 986  P486  6.4254144
#> 987  P487  6.6132597
#> 988  P488  9.1712707
#> 989  P489 10.4861878
#> 990  P490 10.6574586
#> 991  P491  8.9723757
#> 992  P492  5.7458564
#> 993  P493  6.2762431
#> 994  P494 10.5027624
#> 995  P495 10.6574586
#> 996  P496 10.6077348
#> 997  P497 10.6574586
#> 998  P498  0.0000000
#> 999  P499 12.5303867
#> 1000 P500  6.4254144